Refactor routes to Blueprint architecture with app factory pattern
- Created app/routes/ folder structure
- Created auth.py Blueprint (register, login, logout routes)
- Created calculate.py Blueprint (calculate route)
- Created history.py Blueprint (history, delete routes)
- Registered all Blueprints in app.py
- Converted to app factory pattern with create_app() function
- Updated all templates to use Blueprint-prefixed url_for() calls (auth.login, calculate.calculate, etc.)
- Added missing imports (url_for, redirect) to Blueprint files
- Moved LoginManager.login_view to point to 'auth.login'
- All features tested and working (registration, login, calculator, history, delete, logout)

Relates to #23, #25
