+++ templates/base.html
@@ -13,11 +13,11 @@
         <h1>Welcome to QuantEngine!!!</h1>
         <h3>Your one-stop solution for investment scenario analysis.</h3>
         <a href="{{url_for('index')}}">Home</a>
-        <a href="{{url_for('register')}}">Register</a>
-        <a href="{{url_for('login')}}">Log In</a>
-        <a href="{{url_for('calculate')}}">Risk & Return</a>
-        <a href="{{url_for('history')}}">History</a>
-        <a href="{{url_for('logout')}}">Log Out</a>
+        <a href="{{url_for('auth.register')}}">Register</a>
+        <a href="{{url_for('auth.login')}}">Log In</a>
+        <a href="{{url_for('calculate.calculate')}}">Risk & Return</a>
+        <a href="{{url_for('history.history')}}">History</a>
+        <a href="{{url_for('auth.logout')}}">Log Out</a>
     </header>
 
     <main>
